Add new features and fix bugs #4

+++ app/Http/Controllers/CatagoriesController.php
@@ -15,17 +15,20 @@ class CatagoriesController extends Controller
     // display a list of categories in the database
     public function index(): View
     {
+        $this->authorize('view category module');
         return view('pages.catagories.index');
     }
 
     // displays the page for the user to add catagories
     public function create(): View
     {
+        $this->authorize('add category');
         return view('pages.catagories.create');
     }
     // view sigle record
     public function show(Catagory $category): View
     {
+        $this->authorize('view category');
         return view('pages.catagories.show', compact('category'));
     }
 
@@ -39,6 +42,7 @@ class CatagoriesController extends Controller
     // display the edit page
     public function edit(Catagory $category): View
     {
+        $this->authorize('edit category');
         return view('pages.catagories.edit', compact('category'));
     }
     // update record
@@ -52,7 +56,11 @@ class CatagoriesController extends Controller
     // delete record
     public function destroy(Catagory $category): RedirectResponse
     {
+        // allow to delete resource
+        $this->authorize('delete category');
+        // delete resource
         $category->delete();
+        // go back to index page
         return to_route('catagories.index')->with('success', 'Category has been deleted!');
     }
 }
